+++ tests/projects/c++/protobuf/src/main.cpp
@@ -1,0 +1,10 @@
+#include <iostream>
+#include "test.pb.h"
+
+using namespace std;
+
+int main(int argc, char** argv)
+{
+    cout << "hello world!" << endl; 
+    return 0;
+}