+++ xmake/actions/require/impl/action/patch.lua
@@ -30,7 +30,7 @@ function _patch(package, patch_url, patch_hash)
     vprint("patching %s to %s-%s ..", patch_url, package:name(), package:version_str())
  
     -- get the patch file
-    local patch_file = path.join(os.tmpdir(), "patches", package:name(), package:version_str(), path.filename(patch_url))
+    local patch_file = path.join(os.tmpdir(), "patches", package:name(), package:version_str(), (path.filename(patch_url):gsub("%?.+$", "")))
 
     -- the package file have been downloaded?
     local cached = true